Separate preset theme picker from color scheme picker & utils
This way when there is a forced theme we can include one but not the other. This fixes the dark mode in the theme preview.
